requirements.txt
flask
feedparser
gspread
google-auth